asn1: Template backend partial support for SETs
This adds enough support for SET { ... } types to the template backend
to allow the X.690 sample test to be run with the template backend.

Limitations:

 - For DER encoding, the SET { ... } members must be manually sorted by
   the module author.

 - Decoding of out-of-order encodings (BER) is not supported at this
   time.

These shortcomings will be addressed later.

Note that because we don't parse IMPORTed modules at this time, we can't
sort SET { ... } members at compile time if any of them out an
outer-most tag that the compiler cannot see without learning to parse
IMPORTed modules.