kdc: allow TGS requests from synthetic clients
Previously the TGS allowed synthetic clients (that is, clients that do not
exist in the HDB) only if PKINIT was used. Since the introduction of support
for GSS-API pre-authentication, synthetic clients can be used without PKINIT.

To support these in the TGS, use a new authorization data element to indicate
the AS validated the usage of a synthetic client.