kdc: Update PAC hooks for Samba.
Samba needs the PKINIT details to allows PAC_CRENDENTIAL_INFO
to be added to the PAC when using PKINIT. In that case
PAC_CRENDENTIAL_INFO contains an encrypted PAC_CRENDENTIAL_DATA.

BUG: https://bugzilla.samba.org/show_bug.cgi?id=11441

(similar to Samba commit 0022ea9efb0e7809fa2d060b294320eb0479cdd2)

Samba also needs KDC support for PAC_ATTRIBUTES_INFO PAC buffer

This records the client's request for a PAC to be sent (or not sent)
to the target server by means of a bitfield in the PAC, called
PAC_ATTRIBUTES_INFO.  This adds the hooks in Heimdal to support
that.  Because in AD after 2021-11 all TGT tickets have a PAC, the
decision point on calling the generate_pac() is using the former
send_pac_p() is removed and instead the called plugin will determine
that.

BUG: https://bugzilla.samba.org/show_bug.cgi?id=14561

(similar to Samba commit 2f9245f2a549bd89829d7807ec525c54ff61f8e5)

Samba needs the PAC hook to always be called, so we always generate
a PAC for S4U2Self.

If we decided not to put a PAC into the ticket, mspac would be NULL
here, and the resulting ticket would not contain a PAC. This could
happen if there was a request to omit the PAC or the service did not
require authorization data. Ensure that we always generate a PAC.

(Similar to samba commit 1f4f3018c5001b289b91959a72d00575c8fc0ac1)

Allow Samba to not generate extra PAC buffers for S4U2Self service ticket

When generating a service ticket for S4U2Self, we want to avoid adding
the additional PAC_ATTRIBUTES_INFO and PAC_REQUESTER_SID buffers.

The Heimdal change is to pass in the extra "server" record to determine
if these buffers should be omitted

(Similar to Samba commit 9bd26804852d957f81cb311e5142f9190f9afa65)

Finally

Original authors:
 - Joseph Sutton <josephsutton@catalyst.net.nz>
 - Andrew Bartlett <abartlet@samba.org>
 - Stefan Metzmacher <metze@samba.org>